Invalidate pooled sessions for the same destination when we got a KeepAliveDisconnected exception. closes #30.
I know the original intent of #30 is 'create a new (non cached) session
when we got a KeepAliveDisconnected' but the suggested fix could create
too much session to the same site I guess... So at this time, I decided
to invalidate all sessions to the same site. It must include live
sessions (not expired) but it should not cause performance degrade in
the situation such as we got KeepAliveDisconnected.