Checks only authenticators that are set
We want the ability to only set certain kinds of authenticators if we
already know what they will accept.

Explanation:
1) We know that one of our remote systems *anly* accepts BasicAuth
2) We have limited, expensive bandwidth and our messages always contain a
large attachment
3) If we use the standard HttpClient, which goes through the whole gamut of
authenticators... it will continually send this large attachment multiple
times... using authenticators that we *already know* will be refused, thus
wasting our bandwidth and time.

Therefore, we want to restrict the authenticator set to only the known
authenticators... when we *already know* the authentication schemes that the
remote system accepts.

The way I'e done this is quite simple - and the ddefault behaviour of the
system will remain the same.

When you "set" the authenticator, it marks it as usable.
If you use the default "set_auth" method - it will mark all authenticators
as usable, just as it curently does.

*but* if you already know the authentication method... then only set the
methods you will use...

For example, if you know that it will only accept basic auth, and negotiate
auth, use:

@client = HTTPClient::WWWauth.new
@client.basic_auth.set_auth('myuser','mypass')
@client.negotiate_auth.set_auth('myuser','mypass')

More-complicated authenticators may require other params, or even no params
eg the following doesn't actually *need* the given params...
@client.sspi_negotiate_auth.set_auth('uri.com','myuser','mypass')

This change frees us up to use the nice fallback negotiation if we need it,
but does not force us to go through the waterfall process if it's *not*
needed.

Having options is a Good Thing.