Optimize cached session access performance
Change @sess_pool from an Array to a Hash with Site as a key.  Each
value is an Array of sessions.  This should improve cache access
performance if multiple sites are in cache.

In addition to the change, do scrub the cache every @keep_alive_timeout
seconds.  This also should improve cache access performance since it was
scrubbing the cache every time.  See #69