Make dump_file and dump_chunks work with CompositeReadIO
Not sure how to go about explaining this in a short way.

This change was necessary to support streaming file uploads in Faraday
with HTTPClient. Its CompositeReadIO io object was not working with the
previous implementation, where you pass buffer as second argument.

Tests on your end are also passing, although I'm not sure if you have
file uploads covered.