Add CLI
bin/httpclient is installed as CLI.

Usage: 1) % httpclient get https://www.google.co.jp/ q=ruby
Usage: 2) %httpclient

For 1) it issues a GET request to the given URI and shows the wiredump and
the parsed result.  For 2) it invokes irb shell with the binding that has a
HTTPClient as 'self'.  You can call HTTPClient instance methods like;

  > get "https://www.google.co.jp/", :q => :ruby