Change default SSL version to nil to use version negotiation.
Old default SSLv3 disables version negotiation and does not
connect to TSL-only servers. This was really problematic when
httpclient was not used directly but by third-party libraries.

This commit changes the default version to nil that will not
set a SSL version on context creation. This way client and server
can negotiation the SSL version and choose the most secure
available. SSLv2 is still disabled in the cipher suite setting
with "!SSLv2".