Allow post and put accepts :query parameter
Code looks weird but needed. To keep compatibility,
  post('http://localhost/', :query => query)
is a POST method WITHOUT query and with "query=query" BODY. Now
HTTPClient accepts :query parameter if :body parameter given with it.
  post('http://localhost/', :query => query, :body => body)
is a POST method to http://localhost/?query with 'body' as message-body.
Closes #83.

Also this commit simplifies argument handling so that we can keep
post/post_content/post_async in sync.