Keep quoted cookie value as is and send back to the server
Try to send back the given Cookie value to the server as is to the
extent possible. When Set-Cookie is 'key="value"', send 'key="value"' to
the server not 'key=value'.

Implementation: keep Cookie values without normalizing internally and do
normalize when Cookie#value is called. The raw_value is written to
Cookie file as well.  Closes #145.