Add strict_response_size_check option
An option to raise BadResponse if response size does not match with
Content-Length header in response.

Currently HTTPClient ignores response size even if the size does not
match.  Perhaps because of bad HTTP servers, browsers ignore EOF even for
file download (simple GET.).

https://code.google.com/p/chromium/codesearch#chromium/src/content/browser/download/download_request_core.cc&q=ERR_INCOMPLETE_CHUNKED_ENCODING&sq=package:chromium&type=cs&l=534
Chromium seems to ignore incomplete download (both for
ERR_CONTENT_LENGTH_MISMATCH and ERR_INCOMPLETE_CHUNKED_ENCODING) for
download.

https://bugzilla.mozilla.org/show_bug.cgi?id=237623
The comment in this issue says the newer Firefox detects EOF but I
confirmed that Firefox also ignores EOF.

httpclient is not a browser so it should not suppress EOFError but I
also worry about incompatible behavior with previous versions. So this
commit adds an option but it's disabled by default.  In the next version
update I'll enable this.