Set SSL connect timeout for unproxied connections
As reported in #341, SSL connections are not timing out correctly. This
fixes that problem for unproxied SSL connections -- I haven't dove into
how to do this for proxied SSL connections yet.

As noted in that issue, we don't have separate timeouts for connect vs
send/receive, so I use the singular timeout as connect timeout.