Don't leave invalid sessions around if closing socket fails
We use https://github.com/googleapis/google-cloud-ruby to connect to BigQuery from AWS. We have an AWS NAT in place that kills idle connections after 5 minutes. When the previous code tried to clean up old sessions, it would fail with `Broken pipe` on `sess.close`... and the invalid sessions would remain in the cache (because `pool.replace` was never called).

So rescuing from `HTTPClient::KeepAliveDisconnected` to retry it would immediately fail again, because it would use the invalid session that wasn't removed.